+++ packages/browser/src/node/state.ts
@@ -1,5 +1,5 @@
 import type { BrowserServerState as IBrowserServerState } from 'vitest/node'
-import type { WebSocketBrowserRPC } from './types'
+import type { WebSocketBrowserRPC } from '../types'
 
 export class BrowserServerState implements IBrowserServerState {
   public readonly orchestrators: Map<string, WebSocketBrowserRPC> = new Map()
