+++ packages/vitest/src/integrations/vi.ts
@@ -15,6 +15,7 @@ import { getWorkerState, isChildProcess, resetModules, waitForImportsToResolve }
 import { parseSingleStack } from '../utils/source-map'
 import { FakeTimers } from './mock/timers'
 import { waitFor, waitUntil } from './wait'
+import { ModuleCacheMap } from 'vite-node'
 
 type ESModuleExports = Record<string, unknown>
 
@@ -689,7 +690,13 @@ function createVitest(): VitestUtils {
     },
 
     resetModules() {
-      resetModules(workerState.moduleCache)
+      // @ts-expect-error injected by the browser
+      if (typeof __vitest_browser_runner__ !== 'undefined') {
+        throw new Error(
+          `vi.resetModules() is not supported in the browser environment.`
+        )
+      }
+      resetModules(workerState.moduleCache as ModuleCacheMap)
       return utils
     },
