+++ packages/vitest/src/node/types/worker.ts
@@ -1,5 +1,5 @@
 import type { MessagePort } from 'node:worker_threads'
-import type { ModuleCacheMap, ModuleExecutionInfo } from 'vite-node'
+import type { ModuleExecutionInfo } from 'vite-node'
 import type { ContextRPC } from '../../types/worker'
 
 export interface WorkerContext extends ContextRPC {
@@ -8,7 +8,6 @@ export interface WorkerContext extends ContextRPC {
 
 declare module 'vitest' {
   interface WorkerGlobalState {
-    moduleCache: ModuleCacheMap
     moduleExecutionInfo?: ModuleExecutionInfo
   }
 }
