+++ packages/vitest/src/public/browser.ts
@@ -1,5 +1,4 @@
 export {
-  getCoverageProvider,
   startCoverageInsideWorker,
   stopCoverageInsideWorker,
   takeCoverageInsideWorker,
