+++ packages/vitest/src/public/coverage.ts
@@ -1,1 +1,1 @@
-export { BaseCoverageProvider } from '../utils/coverage'
+export { BaseCoverageProvider } from '../node/coverage'
