+++ packages/vitest/src/runtime/runners/test.ts
@@ -9,7 +9,7 @@ import type {
   VitestRunnerImportSource,
 } from '@vitest/runner'
 import type { SerializedConfig } from '../config'
-import type { VitestExecutor } from '../execute'
+// import type { VitestExecutor } from '../execute'
 import { getState, GLOBAL_EXPECT, setState } from '@vitest/expect'
 import { getNames, getTestName, getTests } from '@vitest/runner/utils'
 import { createExpect } from '../../integrations/chai/index'
@@ -22,7 +22,7 @@ import { getWorkerState } from '../utils'
 export class VitestTestRunner implements VitestRunner {
   private snapshotClient = getSnapshotClient()
   private workerState = getWorkerState()
-  private __vitest_executor!: VitestExecutor
+  private __vitest_executor!: any
   private cancelRun = false
 
   private assertionsErrors = new WeakMap<Readonly<Task>, Error>()
