+++ tsconfig.base.json
@@ -24,17 +24,9 @@
       "~/*": ["./packages/ui/client/*"],
       "vitest": ["./packages/vitest/src/public/index.ts"],
       "vitest/globals": ["./packages/vitest/globals.d.ts"],
-      "vitest/node": ["./packages/vitest/src/public/node.ts"],
-      "vitest/reporters": ["./packages/vitest/src/public/reporters.ts"],
-      "vitest/execute": ["./packages/vitest/src/public/execute.ts"],
-      "vitest/config": ["./packages/vitest/src/public/config.ts"],
-      "vitest/coverage": ["./packages/vitest/src/public/coverage.ts"],
-      "vitest/browser": ["./packages/vitest/src/public/browser.ts"],
-      "vitest/runners": ["./packages/vitest/src/public/runners.ts"],
+      "vitest/*": ["./packages/vitest/src/public/*"],
       "vite-node": ["./packages/vite-node/src/index.ts"],
-      "vite-node/client": ["./packages/vite-node/src/client.ts"],
-      "vite-node/server": ["./packages/vite-node/src/server.ts"],
-      "vite-node/utils": ["./packages/vite-node/src/utils.ts"]
+      "vite-node/*": ["./packages/vite-node/src/*"]
     },
     "strict": true,
     "declaration": true,
